Sprites~SCSS~NavBar
I've completed the elements in the nav bar and now need to just lay it out. Learning SCSS as I go and it's very useful. Learned how to use a sprite as a background image and use sizing to pick which sprite to show. https://getspritexy.netlify.app/ is the goat.